Fixes #846 Use import path instead of .
