+++ src/util.ts
@@ -5,7 +5,7 @@
 
 import vscode = require('vscode');
 import path = require('path');
-import { getGoRuntimePath, getBinPathWithPreferredGopath, resolvePath} from './goPath';
+import { getGoRuntimePath, getBinPathWithPreferredGopath, resolvePath } from './goPath';
 import cp = require('child_process');
 import TelemetryReporter from 'vscode-extension-telemetry';
 
@@ -159,7 +159,7 @@ export function getGoVersion(): Promise<SemVersion> {
 	}
 
 	if (goVersion) {
-		sendTelemetryEvent('getGoVersion', {version: `${goVersion.major}.${goVersion.minor}`});
+		sendTelemetryEvent('getGoVersion', { version: `${goVersion.major}.${goVersion.minor}` });
 		return Promise.resolve(goVersion);
 	}
 	return new Promise<SemVersion>((resolve, reject) => {
@@ -170,9 +170,9 @@ export function getGoVersion(): Promise<SemVersion> {
 					major: parseInt(matches[1]),
 					minor: parseInt(matches[2])
 				};
-				sendTelemetryEvent('getGoVersion', {version: `${goVersion.major}.${goVersion.minor}`});
+				sendTelemetryEvent('getGoVersion', { version: `${goVersion.major}.${goVersion.minor}` });
 			} else {
-				sendTelemetryEvent('getGoVersion', {version: stdout});
+				sendTelemetryEvent('getGoVersion', { version: stdout });
 			}
 			return resolve(goVersion);
 		});
@@ -225,13 +225,13 @@ export function isGoPathSet(): boolean {
 }
 
 export function sendTelemetryEvent(eventName: string, properties?: {
-			[key: string]: string;
-		}, measures?: {
-			[key: string]: number;
-		}): void {
+	[key: string]: string;
+}, measures?: {
+	[key: string]: number;
+}): void {
 
-			telemtryReporter = telemtryReporter ? telemtryReporter : new TelemetryReporter(extensionId, extensionVersion, aiKey);
-			telemtryReporter.sendTelemetryEvent(eventName, properties, measures);
+	telemtryReporter = telemtryReporter ? telemtryReporter : new TelemetryReporter(extensionId, extensionVersion, aiKey);
+	telemtryReporter.sendTelemetryEvent(eventName, properties, measures);
 }
 
 export function isPositionInString(document: vscode.TextDocument, position: vscode.Position): boolean {
@@ -257,3 +257,27 @@ export function getBinPath(tool: string): string {
 	return getBinPathWithPreferredGopath(tool, getToolsGopath());
 }
 
+export function getCurrentGoWorkspaceFromGOPATH(currentFileDirPath: string): string {
+	let workspaces: string[] = process.env['GOPATH'].split(path.delimiter);
+	let currentWorkspace = path.join(workspaces[0], 'src');
+
+	// Workaround for issue in https://github.com/Microsoft/vscode/issues/9448#issuecomment-244804026
+	if (process.platform === 'win32') {
+		currentFileDirPath = currentFileDirPath.substr(0, 1).toUpperCase() + currentFileDirPath.substr(1);
+	}
+
+	// In case of multiple workspaces, find current workspace by checking if current file is
+	// under any of the workspaces in $GOPATH
+	for (let i = 1; i < workspaces.length; i++) {
+		let possibleCurrentWorkspace = path.join(workspaces[i], 'src');
+		if (currentFileDirPath.startsWith(possibleCurrentWorkspace)) {
+			// In case of nested workspaces, (example: both /Users/me and /Users/me/src/a/b/c are in $GOPATH)
+			// both parent & child workspace in the nested workspaces pair can make it inside the above if block
+			// Therefore, the below check will take longer (more specific to current file) of the two
+			if (possibleCurrentWorkspace.length > currentWorkspace.length) {
+				currentWorkspace = possibleCurrentWorkspace;
+			}
+		}
+	}
+	return currentWorkspace;
+}
